backend/pyproject.toml
[tool.poetry]
name = "crucible-backend"
version = "1.0.0"
description = "Backend API for Crucible crossword puzzle platform"
authors = ["Crucible Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
sqlalchemy = "^2.0.30"
psycopg2-binary = "^2.9.9"
pydantic = {extras = ["email"], version = "^2.7.1"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"